+++ timeforge-api/src/main/java/br/com/timeforge/timeforge_api/service/TurmaService.java
@@ -3,6 +3,8 @@ package br.com.timeforge.timeforge_api.service;
 import br.com.timeforge.timeforge_api.dto.request.TurmaRequestDTO;
 import br.com.timeforge.timeforge_api.dto.response.TurmaResponseDTO;
 import br.com.timeforge.timeforge_api.entity.Turma;
+import br.com.timeforge.timeforge_api.repository.AulaRepository;
+import br.com.timeforge.timeforge_api.repository.TurmaDisciplinaRepository;
 import br.com.timeforge.timeforge_api.repository.TurmaRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
@@ -15,9 +17,17 @@ import java.util.stream.Collectors;
 public class TurmaService {
 
   private final TurmaRepository repository;
+  private final TurmaDisciplinaRepository turmaDisciplinaRepository;
+  private final AulaRepository aulaRepository;
 
-  public TurmaService(TurmaRepository repository) {
+  public TurmaService(
+          TurmaRepository repository,
+          TurmaDisciplinaRepository turmaDisciplinaRepository,
+          AulaRepository aulaRepository
+  ) {
     this.repository = repository;
+    this.turmaDisciplinaRepository = turmaDisciplinaRepository;
+    this.aulaRepository = aulaRepository;
   }
 
   public List<TurmaResponseDTO> listarTurmas() {
@@ -37,7 +47,6 @@ public class TurmaService {
   public TurmaResponseDTO cadastrarTurma(TurmaRequestDTO turmaObject) {
     Turma turmaSalva = toEntity(turmaObject);
     turmaSalva = repository.save(turmaSalva);
-
     return toResponseDTO(turmaSalva);
   }
 
@@ -55,6 +64,20 @@ public class TurmaService {
   public void deletarTurma(Long id) {
     Turma turmaEncontrada = repository.findById(id)
             .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, "Turma com id (" + id + ") nao encontrada!"));
+
+    if (turmaDisciplinaRepository.existsByTurmaId(id)) {
+      throw new ResponseStatusException(
+              HttpStatus.CONFLICT,
+              "Nao e possivel excluir turma com id (" + id + "), pois existem vinculacoes em turma_disciplina."
+      );
+    }
+
+    if (aulaRepository.existsByTurmaId(id)) {
+      throw new ResponseStatusException(
+              HttpStatus.CONFLICT,
+              "Nao e possivel excluir turma com id (" + id + "), pois existem aulas vinculadas."
+      );
+    }
 
     repository.delete(turmaEncontrada);
   }
